CRA doesn't like symlinks, but backend can handle it
